Many Many Changes
The main thing is that the main home page of the application's code was
changed to be better. You can now see that everything is handled by
various subsystems now. This means that #13 is resolved.

As well, the placeholder for end game logic has been added, just need to
confirm the logic for when to move on to the end game.

There were other various changes and performance enhancements made.
Tested to be in usable condition, will need to test further before the
actual reindeer hunt.