Fixed performance and de-enhancing bugs
There was a bug where closing the student or match datacard window caused search filters to reset. Was fixed.

Also made it so that all subthreads close now on exit, by setting them to background.

Also made it so that the progress bar actually updates with progress, which was a simple division by two integers always making 0.

Also added some statistics parameters to the home screen, namely the number of students passed and the total number of students, in or out.

Tested the application a little more. Seems stable.

Updated installer.