Resolved exporting bug, and added more feedback to result import - V1.0.17
V1.0.15 - Made it so that importing students after round 0 is possible. They will simply be matched for in the round after the one where they were imported.

V1.0.16 - Fixed a bug where when exporting data, if the export file existed already, the program would crash.

V1.0.17 - Added more user feedback to match result importing. Specifically, a log file is now created with ALL the problems encountered during the import, so the user can resolve all the problems in one go.