QR code Addition! - V1.1.22
V1.1 - New Minor build because of the large QR code feature addition.

Build 20 - The QR feature. When enabled, instant print puts a QR code with a link to the submission form autofilled with the student's information.

Build 21 - The ability to change the submission form's url.
Build 22 - The ability to reset the submission form's url, and disable QR code printing.

Has been tested compatible with the datafile  from versions after 1.0.0.