Match Editing Tools - V1.2.42 (Huge Changes)
V1.1.23 - Added the tools menu item, and the ability to add matches to the match edit queue, which is the same as the student pass queue.

V1.1.24 - Added the Close All Matches Tool, which closes all the currently open matches, eliminating the students in them.

V1.1.25 - Added the Close Selected Matches tool, which closes the matches selected in the match edit queue.

V1.1.26 - Added the Print Selected Matches tool, which instant prints licenses for matches in the edit matches queue.

V1.2.27 - V1.2.39 - Added the match edit tool, which gives the user the ability to edit matches, and put students back in the hunt.

V1.2.40 - Improved the compatibility checker, and added a version number key to the DataFile to make checking quicker.

V1.2.41 - Added a static MakeMatches method to the Matcher for the ability for other objects to use the method to generate matches randomly the same way the matcher would.

V1.2.42 - Various small bug fixes, improvements, changes.