Completed implementation of no student id
The application can now be fully used with empty student IDs.

The students who don't have IDs are given an ID that is randomly
generated at their import time (with a special prefix so that you know
that it's been randomly generated) and works this way.

Therefore, #69 has been resolved.