Prettify written data
The data written to the data file is now json prettified automatically
as it is written, which makes it much easier for humans to edit and
stuff.

Also did a couple of small improvements here and there where there were
unecessary namespace qualifyings for some types and such.

Also renamed a stream which had the same name as another stream because
that's bad.