Began update user manual
The user manual needs updating (as specified in #68) so I began this
process by changing the version number included in the manual and by
re-organizing it a little bit.

The user manual will be primarily action-focused rather than explaining
what each of the UI elements do, but the old explanations of each UI
element will still exist in its own section.