Made quick-fix to printer's sorting
The sorting algorithm in the instant printer was using the student's
grade and not homeroom number to sort through the licenses, which is bad
since it means that a student who is in a different grade than their
homeroom can cause empty pages of licenses to be printed. This has been
resolved in a not very clean way however.

Todo has been placed but for now this works.