Fixed closing matches async
The event called at the end of closing matches asynchronously was
causing issues with accessing GUI elements, which then caused the
application to crash since GUI elements shouldn't be modified from a
worker thread. Therefore, I put together a bad solution of using the
Dispatcher to call the event from the main thread.